simplify layout
