+++ tests/test_sync_validation.py
@@ -3,7 +3,7 @@ import pytest
 
 from cent import (
     Client,
-    CentAPIError,
+    CentResponseError,
     PublishRequest,
     BroadcastRequest,
     PresenceRequest,
@@ -155,7 +155,7 @@ def test_batch(sync_client: Client) -> None:
 
 
 def test_error_publish(sync_client: Client) -> None:
-    with pytest.raises(CentAPIError, match="unknown channel") as exc_info:
+    with pytest.raises(CentResponseError, match="unknown channel") as exc_info:
         sync_client.publish(
             "undefined_channel:123",
             {"data": "data"},
